metadata-api/requirements.txt
beautifulsoup4==4.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
google==2.0.2
google-api-core==1.14.2
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-storage==1.19.1
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
protobuf==3.9.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.3
toolz==0.10.0
urllib3==1.25.6
Werkzeug==0.16.0
